+++ changelog.md
@@ -5,7 +5,7 @@
 - Catch errors during pings to Google Analytics, for example in case of a
   missing internet connection
 - Track additional browser data, such as screen size and language
-- Added tests for usage running in a dart:html context
+- Added tests for `usage` running in a dart:html context
 -
 
 ## 0.0.4